# specs/012-order-bom-consumption/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: BOM & Stock Consumption (spec 012)
  version: 1.0.0
  description: |
    REST endpoints добавляемые/расширяемые спекой 012 в модуле `inventory`.
    Не описывает существующие эндпоинты материалов и приходов из спеки 011.

servers:
  - url: http://localhost:8080
    description: Local dev (Docker compose)

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    BigDecimalString:
      type: string
      pattern: '^[0-9]+(\.[0-9]{1,4})?$'
      description: BigDecimal со scale ≤ 4. Положительное.
      example: "10.0000"

    MeasurementUnit:
      type: string
      enum: [PIECE, KILOGRAM, METER, LITER, SQUARE_METER, CUBIC_METER]

    OrderStatus:
      type: string
      enum: [NEW, IN_WORK, READY, SHIPPED]

    InventoryError:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
        field: { type: string, nullable: true }
        available:
          $ref: '#/components/schemas/BigDecimalString'

    BomLine:
      type: object
      required:
        [id, orderId, materialId, materialName, materialUnit, quantity, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        materialId: { type: string, format: uuid }
        materialName: { type: string }
        materialUnit: { $ref: '#/components/schemas/MeasurementUnit' }
        quantity: { $ref: '#/components/schemas/BigDecimalString' }
        comment: { type: string, nullable: true, maxLength: 500 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    BomLineList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/BomLine' }

    BomLineCreateRequest:
      type: object
      required: [materialId, quantity]
      properties:
        materialId: { type: string, format: uuid }
        quantity: { $ref: '#/components/schemas/BigDecimalString' }
        comment: { type: string, nullable: true, maxLength: 500 }

    BomLineUpdateRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { $ref: '#/components/schemas/BigDecimalString' }
        comment: { type: string, nullable: true, maxLength: 500 }

    ConsumeRequest:
      type: object
      required: [orderId, quantity]
      properties:
        orderId: { type: string, format: uuid }
        quantity: { $ref: '#/components/schemas/BigDecimalString' }
        comment: { type: string, nullable: true, maxLength: 500 }

    StockMovement:
      type: object
      required: [id, materialId, materialName, materialUnit, movementType, quantity, actorDisplayName, createdAt]
      properties:
        id: { type: string, format: uuid }
        materialId: { type: string, format: uuid }
        materialName: { type: string }
        materialUnit: { $ref: '#/components/schemas/MeasurementUnit' }
        movementType:
          type: string
          enum: [RECEIPT, CONSUMPTION]
        orderId:
          type: string
          format: uuid
          nullable: true
          description: Обязателен для CONSUMPTION, всегда null для RECEIPT
        orderNumber:
          type: string
          nullable: true
        quantity: { $ref: '#/components/schemas/BigDecimalString' }
        comment: { type: string, nullable: true }
        actorDisplayName: { type: string }
        createdAt: { type: string, format: date-time }

    MaterialUsageRow:
      type: object
      required:
        [materialId, materialName, materialUnit, requiredQuantity, consumedQuantity, remainingToConsume, overconsumption]
      properties:
        materialId: { type: string, format: uuid }
        materialName: { type: string }
        materialUnit: { $ref: '#/components/schemas/MeasurementUnit' }
        requiredQuantity: { $ref: '#/components/schemas/BigDecimalString' }
        consumedQuantity: { $ref: '#/components/schemas/BigDecimalString' }
        remainingToConsume: { $ref: '#/components/schemas/BigDecimalString' }
        overconsumption: { $ref: '#/components/schemas/BigDecimalString' }

    MaterialUsage:
      type: object
      required: [orderId, rows]
      properties:
        orderId: { type: string, format: uuid }
        rows:
          type: array
          items: { $ref: '#/components/schemas/MaterialUsageRow' }

    InventoryOrderSummary:
      type: object
      required: [id, orderNumber, customerName, status]
      properties:
        id: { type: string, format: uuid }
        orderNumber: { type: string }
        customerName: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }

    InventoryOrderList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/InventoryOrderSummary' }

paths:
  /api/orders/{orderId}/bom:
    parameters:
      - name: orderId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: List BOM lines for order
      tags: [BOM]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BomLineList' }
        '401': { description: Unauthorized }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
    post:
      summary: Add BOM line
      tags: [BOM]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BomLineCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BomLine' }
        '400':
          description: Validation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '404':
          description: Order or material not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '409':
          description: ORDER_LOCKED or BOM_LINE_DUPLICATE
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }

  /api/orders/{orderId}/bom/{lineId}:
    parameters:
      - name: orderId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: lineId
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      summary: Update BOM line (quantity / comment)
      tags: [BOM]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BomLineUpdateRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BomLine' }
        '400':
          description: Validation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '409':
          description: ORDER_LOCKED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
    delete:
      summary: Remove BOM line
      tags: [BOM]
      responses:
        '204': { description: Removed }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '409':
          description: ORDER_LOCKED or BOM_LINE_HAS_CONSUMPTION
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }

  /api/materials/{id}/consume:
    post:
      summary: Consume stock (with order reference)
      tags: [Consumption]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConsumeRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StockMovement' }
        '400':
          description: Validation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '404':
          description: Material or order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '409':
          description: ORDER_LOCKED, MATERIAL_NOT_IN_BOM or INSUFFICIENT_STOCK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }

  /api/orders/{orderId}/material-usage:
    get:
      summary: Aggregated material usage per order (BOM vs consumption)
      tags: [Usage]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MaterialUsage' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }

  /api/orders/active-for-consumption:
    get:
      summary: List active orders eligible for stock consumption (have BOM, not SHIPPED)
      description: Тонкий фасад над OrderLookupPort.searchActiveOrdersForConsumption — для picker в StockConsumeDialog (US2 alt-entry).
      tags: [Consumption]
      parameters:
        - name: search
          in: query
          schema: { type: string }
          description: Подстрока по order_number или customer_name (case-insensitive)
        - name: limit
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryOrderList' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InventoryError' }